1. Split out state_dict related hooks to a seperate file to reduce development conflicts. 2. Remove unused APIs.
[ghstack-poisoned]
